Add support for scoped entities
