+++ src/index.js
@@ -1,3 +1,6 @@
-export { default as makeEntity, createEntity } from './makeEntity';
+export { default as createEntity } from './createEntity';
+export { default as makeEntity } from './makeEntity';
+export { default as EntityScope } from './EntityScope';
+export { default as useEntity } from './useEntity';
 export { default as useEntityBoundary } from './useEntityBoundary';
 export { selectAll, selectNone, strictEqual, shallowEqual } from './utils';
